[newUI] IDEA-294233 Align Project and Git Widget popups
This solution is far from perfect because ideally it should
be no component's business to calculate insets, as I did it
in ToolbarComboWidget. However, since we're coordinating two
different components, it's not easy to delegate this to their
UIs.

Maybe we could at least query insets from the UI delegates
instead of fetching them from the theme directly, but this
isn't easy either as it would increase coupling between the
component and its UI, as now the component doesn't even
have any idea what the UI delegate's type is, just as it
should be.

The popup of the Run Widget is not aligned yet, this is WIP.

GitOrigin-RevId: f66502943f12c6ef9ff431cc4337da2ee89762e0