[newUI] IDEA-294233 Align Run Widget popup
Similar to Project and Git widgets, but since this one doesn't
extend ToolbarComboWidget, I had to duplicate some code. This is,
of course, not a very good thing, but it will take some refactoring
to avoid it.

Unlike ToolbarComboWidget, this time the widget's insets are retrieved
from its border. This is a good thing and a bad thing: on one hand,
we don't have to query the UIManager directly, on the other hand,
we rely on how it's implemented: if it isn't done using a border,
but, say, is implemented directly in the painting code, it will no
longer work.

GitOrigin-RevId: a05c8828ecd63243d1c2fcaf9725e5839bb7282b