[platform] IDEA-300305 Fix not working buttons in Build Output
Similarly to IDEA-300530 (the issue number similarity is pure coincidence),
the reason for buttons not working was an incorrectly set preferable
focusable component.

To be specific, the preferable focusable component was not updated
when another build view was selected in the build view list. Then
it caused the isShowing() check to fail.

Fixed by extracting the relevant code into a private method to avoid
duplication and then putting the correct logic into it.

GitOrigin-RevId: 50b61d31cc6871db9d589f25e9e4bf241645f153