WI-67842: override InplaceRefactoring.checkRangeContainsOffset() for PHP to proper handling injections in blade files
The platform code calculates a host relative range of injected element and then shifts this range by the injected element offset. That is wrong (at least for Blade). Instead, we need to calculate an element's host offset first and shift its relative range by the calculated offset. However, these changes weren't made to the platform because of unclear and unpredictable impact on other clients (see discussion in IJ-MR-94847).

GitOrigin-RevId: db55e03643635a581cf54ca81ae50128473b0221