[vcs-log] introduce a new method for navigation to the branch head by branch name
VcsLog.jumpToReference method was designed with "Go To Hash/Branch/Tag" action in mind. Branches dashboard should use a more suitable method. This commit introduces a function for navigating only by branch names without searching hashes or tags. This method also handles the situation when the branch is not found due to log not being up-to-date (jumpToReference did not need to do it).

IDEA-301534

GitOrigin-RevId: bfab67e41b26f2618aa1f339b93fb1629f6c1a22