settings-sync: rewrite SettingsSyncPluginManager not to use PersistentStateComponent flow: IDEA-301472
Being a PersistentStateComponent provides the following benefits: automated [de]serialization to/from XML, automated save and load to/from disk. But in the case of the SettingsSyncPluginManager these benefits are not needed: we don't need the state persisted at all (only in the settingsSync/ storage but not in the APP_CONFIG/options). On the other hand, being a PSC complicates the saving/loading flow, especially when we listen to plugin changes.

To simplify the flow, in this commit we stop using the PSC mechanism. Now the SettingsSyncPluginManager is not stored in the APP_CONFIG, only in memory. On the start it collects the information about currently installed plugins, and send a message to the Settings Sync bridge to remember it in the storage and send to other clients if something has changed; if the user installs/enables/disables a plugin, the state also modifies, and a message to the bridge is sent. If something changes by another client, the state gets modified, and the IDE gets modified as well.

Plugins state is serialized as a JSON, both for the server, and in the Git repository.

GitOrigin-RevId: c02721c5f6f47ea0e75a6c707ffe1c288a768c24