[intellij sh] Replace `api` module to `impl`
Actually `impl` module contains all plugin code. Module with direct plugin declaration become as thin as possible, it includes XML file with all extensions from `impl` module. Such solution is needed to avoid circular dependencies with modules for optional plugin's extensions. It's possible to stay with `api` module and include it as dependency, but actually there are a lot of code which should be extracted there, and it was decided to extract all plugin code to such module which actually will be used everywhere. For now `impl` used as dependency in `plugin`, `terminal`, `copyright`, `tests` and much more modules.

GitOrigin-RevId: 677863a8a2920477521673a73eec9ab820c3154f