[kotlin] make KotlinPositionManager and KotlinAlternativeSourceNotificationProvider look for kotlin files in same scopes
Imagine the project where module A depends on library X1.7, and module B depends on library X1.8. When you debug the B module and try to step into sources of the X library, KotlinPositionManager will randomly choose between X1.7 and X1.8. At the same time KotlinAlternativeSourceNotificationProvider will not show the alternative source banner for X1.8, because these libraries would be considered as the 'not in the scope of the debug process' ones. This commit fixes the mentioned behavior by providing the same search scopes for KotlinPositionManager and KotlinAlternativeSourceNotificationProvider.

GitOrigin-RevId: 59cc35e4662be762985adde3a74e4dd26b84645d