[newUI] IDEA-300279 (related) Handle ampersand in Project Widget
ProjectToolbarWidgetAction was setting the text property
using Kotlin property access syntax. However, it defaulted
the mayContainMnemonic argument to true, which lead to &
being incorrectly displayed in the project widget.

Fix by calling the setText() overload with mayContainMnemonic
set to false because I don't think project names may ever contain
mnemonics.

This doesn't fix IDEA-300279, as that ticket is about the Recent
Projects section of the popup, not about the Project Widget itself.
Still, it's related to it, and the fix is very simple, so why bother
creating a separate ticket?

GitOrigin-RevId: f9dfa73a7bb00e4f8a0fd50e8a936b18485c503b