settings-sync: don't lose information about uninstalled plugin: IDEA-301693
Previously we populated the plugins state on the IDE start only by the state of the IDE. However, if a plugin is uninstalled, then it is not known to the IDE anymore, and on restart we lose the information about that uninstalled plugin.

To fix this we pass the latest known state to the SettingsSyncPluginManager to merge it properly with the current state: take the actual information from the IDE for all plugins except for unknown ones (i.e. uninstalled): for those keep the saved information.

In addition to that, remove the obsolete information: e.g. don't store the information about bundled enabled plugin.

GitOrigin-RevId: b5e8e26a0eea44920cb591b2c85fab5f76e2a680