[platform] IDEA-301727 Re-scale icons when scaling is changed
After the previous fix by Aleksey Pivovarov, the icons were
scaled properly according to the system scaling. However,
when the system scaling is changed while IDEA is running,
then icons would re-scale to the usual pixelated mess of
raster scaling instead of redrawing from scratch according
to the new PPI setting.

The core reason behind this was that JBHiDPIScaledImage uses
the JBUIScale.sysScale() method that relies on the provided
GraphicsConfiguration but, if it's not provided, falls back
to the value cached by JBUIScale, which seems to be never updated.

Fix by passing the actual GraphicsConfiguration to ImageUtil and
invalidating the cached image when scaling changes.

GitOrigin-RevId: 415dadca35731ad797a927b505b650ed0dbd2bef