[platform] IDEA-301727 Fix fractional scaling of PNG project icons
Previously an icon loaded from a PNG file would be resized once according
to the user and system scaling levels and then reused and re-scaled,
which looked horribly (until the IDE is restarted).

Fix by reloading and re-scaling the icon from the file if either the
user or system scaling is changed since the last access. Otherwise,
the cached value is used to avoid reloading the file over and
over again.

GitOrigin-RevId: 1229fd4dfe046ec11638b1d65ccd71a30a42a77b