[workspace storage tests] add ignored test which checks equality for entities observed via listener
It's quite common practice to subscribe to events in the workspace model and update entities and associated data in some cache. It's expected that entities stored in the cache will be equal to their counterparts in the current snapshot. It appears that currently dummy modifications break this contract: there will be no events, but the entity in the storage will be changed and become not equal to the old version.

GitOrigin-RevId: 8db0eb8fe4b881efa2d3a201b6864dfe9ff55b94