[platform] IDEA-301727 Defer project icon file loading
Since the actual Icon instance is now created during painting
(as the actual sysScale is only known at that stage),
it's now necessary to avoid disk access at that point.

Also, there was a deferrer for SVG icons wrapped inside another
deferrer, which is ugly and should be unnecessary.

Fix by separating disk operation from the actual scaling.
Now the files are loaded inside the outer deferrer,
and scaling happens later in EDT when painting.

GitOrigin-RevId: f16a713d452ef1eeacd4094e5f417f07aaf823a8