[platform] IDEA-301727 Better workaround for fractional scaling
The core reason of all woes with fractional scaling and incorrect
clipping boils down to how JBR rounds fractional coordinates.
Long story short, translations are rounded 0.5-floor-bias and
clipping is rounded 0.5-ceil-bias, i.e. ceil(x - 0.5) vs floor(x + 0.5).
It results in the image drawn at (1,1) relative to the clipping rectangle
instead of (0,0), that's why the right and bottom parts are clipped.

Work around this by aligning both the clipping rectangle and translation
to integer coordinates before drawing. To do this, use alignTxToInt()
and alignClipToInt() from PaintUtil. The latter needed to be fixed as
it didn't take scaling into account (getClip() and setClip() scale
and unscale the rectangle, leading to even more fractional mess).

For convenience a couple of Kotlin functions were added to simplify this
job at the calling side.

GitOrigin-RevId: 21d39a0b712ec3fa1230fdbc6a82f1485f3af8ce