[platform] IDEA-135035 Restore floating tool window bounds from layout
When the default layout is restored, floating tool windows would not
restore their bounds because the relevant code in FloatingDecorator
only restored the alpha mode.

Fix by restoring the bounds as well.

This doesn't fix the issue completely, as the bounds are not always updated
properly. Another fix is pending.

GitOrigin-RevId: 0396f26a75c001a8d126c0d768f9453671d238af