[platform] IDEA-135035 Update tool window info when moved
When a floating tool window is moved, its floatingBounds
property wasn't updated, which in turn led to the tool
window manager thinking it isn't changed, and therefore
it wasn't restored when the Restore Default Layout action
was performed.

Fix by adding a listener on the floating window decorator.
In order to avoid updating too often, an existing alarm
with 100 ms timeout was used, the same one that already
did the same job in case of window resize (as opposed to
movement). It had to be moved to the class level to reuse it.

GitOrigin-RevId: 30dfba29e87b9c234ba856dd7c7bfec5b409a764