[platform] IDEA-293980 Registry options for independent tool window sizes
Relates to: IDEA-242013 Toolwindow width on the sidebar should be unified.

Add two registry keys:
toolwindow.independent.sizes (for the old UI) and
ide.experimental.ui.toolwindow.independent.sizes (for the new UI).

Both work the same way: if enabled, docked tool windows can be resized independently
(like in the old UI), otherwise they keep the same size as their docking area, switching
between tool windows keeps the same size.

Implementation note: the new UI logic was in ToolWindowManagerImpl.activated(). It turned
out that this function wasn't always called when a tool window is activated. For example,
when the Commit window is opened by clicking on the tool window button, it worked, but
if the Commit window is opened by pressing Ctrl+K, it didn't work, as
ToolWindowManagerImpl.activateToolWindow() was called not via activated() in this case.
Fix this by moving the logic into activateToolWindow().

GitOrigin-RevId: 9902670406669761ded71c2f489501ef962cfbf7