[Kotlin] Avoid NoClassDefFoundError in tests by converting fields to methods
^KTIJ-22768 Fixed

Convert fields to methods in three classes:
* KotlinWithJdkAndRuntimeLightProjectDescriptor
* TestKotlinArtifacts
* ProjectDescriptorWithStdlibSources

This commit breaks `testChangeParameterTypeExpectFunction` test.
Actually, the test was broken before this commit. But it wasn't visible
on CI because if you run all the tests in
`KotlinSuggestedRefactoringTest` then the test is green. It happens
because of the side-effects caused by other tests in the test case.

I will fix `testChangeParameterTypeExpectFunction` in the next commit.

GitOrigin-RevId: 499da37800effac51efe5bf0a6b430231b49fb1a