vcs: scan for vcs roots above content roots top-to-bottom
In the case when there are multiple content roots under the same VCS root,
check it using the outer content root as 'dirToCheckForIgnore'.

The outer folder is less likely to be under '.ignore' list.
We still might skip a valid vcs root with ignored and non-ignored content root siblings,
but checking them for each content root might be too expensive.

GitOrigin-RevId: 18629b2b5372afaef28fef474e7393d56a098b73