[platform] IDEA-242013 Fix wrong tool window width after hide and show
When a tool window is hidden and then shown again, its width would be
calculated from the previously visible window. However, if no window
was previously visible, the relevant code would mistakenly consider
a floating tool window (if any) that was previously docked to the same
area.

This happened because the code only checked the anchor, but not whether
the window was actually docked at the moment, just that it was visible.

This could also happen when switching between tool windows, as the window
is picked randomly, the first that matches the given criteria.

Fix by checking the isDocked property.

GitOrigin-RevId: 2697980716c71ba80ec5d0eaecfb55f5f5741f45