[newUI] IDEA-299064 Update presentations of inline actions
In order to get the presentations updated, inline actions are wrapped
into a temporary ActionGroup in ActionStepBuilder.createInlineActionsItems()
and then expanded in the same way as the regular actions, making use of
the action system.

Then it's just the matter of storing the text and passing it later as
the tooltip.

GitOrigin-RevId: 9a57dfe5dad6ce6271d6319202dfcccf9ef1b47a