Fix KotlinSuggestedRefactoringTest.testChangeParameterTypeExpectFunction test
Related: the previous commit "broke" this test on CI

Alternative fix: Explicitly invoke
`SuggestedRefactoringProviderImpl.getInstance()` at the beginning of
`BaseSuggestedRefactoringTest.doTest`. The alternative fix restricts
potential number of side effects that we need to run the tests. It's
a good and a bad thing at the same time. "restricted number of side
effects" sounds good. The less side effects we have, better the code.

But at the same time, `doHighlighting` has the same number of side
effects as `testHighlighting` (at least I hope so), so it potentialy
eliminates future bugs of the same nature.

GitOrigin-RevId: 5f0422c535c711fd6c78b685df85a08bebb0ac18