[newUI] IDEA-299064 Add tool tips to shortened Run Widget popup actions
If the name was shortened, create a tooltip for the main action
(as opposed to inline actions) with the full run configuration name.

This is done using JComponent.TOOL_TIP_TEXT_KEY already used by
PopupFactoryImpl for tooltips. This involves a deprecated call,
but there seem to be no nice way around it yet, short of adding
a new strongly typed key somewhere.

GitOrigin-RevId: 993ff570b83a294d03a52d66471f72b787af00bd