drop ancient `unregisterKeyboardActions` memory leak workaround
Contrary to the name it leads to exactly same memory leak in KeyboardManager it wants to fight due to ComponentInputMap re-registration (IDEA-296214). It also does not fully do what is expected as only top level maps are cleared.

There's no other way to check if it is still useful but remove it and monitor new memory leaks.

The change also
1. unifies the way `cleanupRootPane` is called (after the full dispose)
2. eliminates double `removeNotify` in `PopupComponent`

GitOrigin-RevId: 6bd50d56e70ca1a9b9f3f4c0080b0af8b601837b