[kotlin] Add custom plugin repositories to GradleMpp[No]JvmRunConfigurationProducersTest
Since 222 platform we have in KotlinGradlePluginVersions.latest version from
`master` or release branch of Kotlin.
In particular, right before release we do have the `1.7.20-233` version there, and this version considered stable. But this version is not published into gradle plugin portal, so we need
to use custom repositories list to get it.

P.s. it is a bit strange, that we are going to start running these tests on the new version right before release, but it is what it is currently

GitOrigin-RevId: 1dbc021e3e19d1e7251b792efd6ea6ffdda3e1b7