settings-sync: don't treat absence of the file on the server as explicit deletion: IDEA-302107
Previously, "remove server data" triggered deletion of the file on the server, and, vice versa, if there were no file on the server, we understood it that it has been explicitly deleted by another client. However, there might be other reasons on why the file could be absent on the server.

In IDEA-302107 there was a race: while one thread was initializing the settings sync, another thread was checking if sync with server is needed. That second thread saw that there is no file on the server, and disabled sync. That issue could be fixed differently, by moving the sync process to the same thread (which might be useful in any case and maybe will be implemented in future). But it is harder to implement, and it is still better to make deletion explicit.

Since this commit, "file is deleted" means that there is a special settings snapshot containing no actual settings, but a single "is deleted" flag.

GitOrigin-RevId: f5348a3de95ee21c4f6c04226589834cca82d1cc