[newUI] IDEA-294733 Use MAIN_MENU instead of MAIN_MENU_IN_POPUP for burger menu
Some actions were missing in the burger menu because it used a different
action place. Since the two menus are supposed to be the same, and there's
A LOT of code that depends on the MAIN_MENU to work correctly, it seems
like the best way to fix it is to ditch MAIN_MENU_IN_POPUP altogether.

This fix didn't work right away, though, because ActionPopupStep expects
the action place to be one of the popup places, which makes sense because
it's a class for popups! Fix that by hacking ActionPopupStep a bit:
now it's allowed to use it in popup places and, additionally, in the main
menu as a special case.

GitOrigin-RevId: dc2ab727aaec2ae55680641fd1cb316595483259