[platform] IDEA-205717 Split image loading and filtering
The main reason with UI freezes when adjusting background image opacity is
the constant unnecessary reloading of the image. Before fixing it, it's better
to refactor the relevant code.

This is part 2. Split image loading and filtering. Maybe we'll need to
cache the unfiltered image later to avoid reloading the image every
time the filter changes. Not sure if it's a good idea, though, as the
image will take a significant chunk of RAM and the filter doesn't
change that often.

GitOrigin-RevId: 4df20b6f444f81f176a7494565e8076e5f4e13d5