[platform] IDEA-205717 Don't reload background image unless necessary
The background image would be reloaded every time any part of the
background image property is changed. However, only the file path
and the flipH/flipV parts are actually used in loading, so,
for example, when the user is moving the opacity slider,
the image keeps reloading over and over again for no good reason,
leading to freezes and OOMs.

Fix that by remembering the combination of the file path and
flipH/flipV settings and not reloading the image unless any
of these are changed.

It's also possible to avoid reloading when the image is flipped,
however, this would require either storing the original image
in the RAM (and it could be huge) or un-flipping the image
before re-flipping it again, which is in theory possible,
but would require deep understanding how the filters work and
whether they are completely lossless and reversible. Since
flipping is not changed that often (unlike the opacity when
the slider is being moved!), it's probably not worth it at this point.

GitOrigin-RevId: 47b9b73024484624d28504d9b79507bf5e96120b