settings-sync: move saving all settings before initializing the setting sync to a better place
This should fix IDEA-301776 when current settings were saved during the initialization, and that led to overwriting of settings applied by the settings sync.

This is an improvement of 53aa0fce: we need to save in all scenarios. Especially, the case IDEA-301776 was not covered in 53aa0fce because enabling settings doesn't invoke the initialize method from SettingsSynchronizer.

GitOrigin-RevId: 726ccf7eb783a2f388f37379fc919ce562387b5e