[newUI] IDEA-294669 Force top-level main menu action groups to be popups
Top-level action groups provided by 3rd party plugins aren't always
marked as popup groups. It wasn't important when the menu was implemented
as a menu bar, but with a popup menu it leads to these actions groups
expanding into individual actions.

Fix by forcibly making these action groups popup groups by setting
isPopupGroup to true for them. Not the best way to do it, but the
best way would be to solve this on the plugins' side which we can't do.

GitOrigin-RevId: 2348ee40f0ec930e806a23cef3687be85320bf3e