[platform] IDEA-302650 MouseDragAwareCaret optimization
To avoid creating a new copy of the MouseEvent, modify its coordinates directly,
and then back. The coordinates don't have setters, so use translation for that.

This also fixes a possible issue with that the super.mouseDragged() can in theory
somehow mutate the copied event (consume it, for example), which wasn't reflected
in the original event.

GitOrigin-RevId: 409da2518774b1d157ac9d2db23074ae109a2b99