[platform] IDEA-302650 Implement Windows and Unix mouse drag selection
On Windows, pretend the mouse is always moving horizontally.

On Unix (Linux and Mac included), if the mouse is moved above or below
the source component, we need to quickly select to the beginning or
the end. We do it by imitating horizontal movement far away from the
component, using coordinates very far away to the left/right.

GitOrigin-RevId: 5379a7f57886d2d2064293a57f0bf35b91092ec3