re-register quick lists as actions after reloading the state from disk: IDEA-303209
Previously, quick lists were reloaded after external changes, but the AnActions correspondent to these lists were not re-registered. These actions are needed to add quick lists to the keymap, without them they are not shown in the keymap, and the shortcut assigned to such a quick list is not working.

GitOrigin-RevId: c2665e029d6de523654480e8f341628c40898717