[newUI] IDEA-294669 Wrap top-level action groups for the burger menu
This is an improvement over the previous solution of the same issue.
Instead of screwing with the template presentation, which may lead
to some unforeseen consequences in some rare cases, wrap top-level
action groups, overriding their update() and making the actual presentation
a popup group.

Kudos to Greg Shrago for coming up with this solution.

GitOrigin-RevId: a585474331e512d96fd94782e001457ab13562d1