IDEA-279398 Fix saving debugger tabs order
Because of removing tabs that was moved outside, there was the shift of positions of the other tabs. And these incorrect positions were saved by debugger tabs logic and in result some tabs become merged after restart.
So, I decided to hide tab that is dragged instead of removing it.

GitOrigin-RevId: ebad284cc1b14d1dc4fd71673c84af878e51f077