[platform] IDEA-93140 Multiple window layouts support
This breaks backwards compatibility with existing configuration files, but as far as I
can see it only leads to one-time resetting layout to the default.

I'm also not very happy with the way Shift+F12 restores the Default layout instead of
the currently selected one. This needs to be fixed, as it completely breaks the usual
"move things around, then reset everything back with Shift+F12" use case.

GitOrigin-RevId: 45d037cd68941f6c14aa70c9ac35c9d860c4e08f