[platform] initial version of WorkspaceFileIndex (IDEA-276394, IDEA-298358)
This introduces WorkspaceFileIndex, which uses information provided by WorkspaceFileIndexContributor extensions to associate files with entities from Workspace Model. The index is updated incrementally when entities from the workspace model are changed, the parts of it taken from AdditionalLibraryRootsProvider and DirectoryIndexExcludePolicy are updated non-incrementally on rootsChanged as before.

Implementations of the existing ProjectFileIndex and ModuleFileIndex are changed to delegate to WorkspaceFileIndex if 'platform.projectModel.workspace.model.file.index' registry option is enabled. Some methods of ProjectFileIndex and ModuleFileIndex still use the old DirectoryIndex, also there is some code which use it directly.

GitOrigin-RevId: 6acb68dd461b372a387354c5234004efca3f2931