[platform] IDEA-93140 Rework UI/UX of multiple layout selection
This probably needs some serious work on it, but at least it's both
usable and convenient now. Store/Restore Default Layout action now
works on the active layout, which means one can easily restore the
previously selected layout with Shift+F12, not the default one.

The same change is made to the ToolWindowDefaultLayoutManager.setLayout()
overload without the name parameter. It now only exists for API compatibility
reasons and whoever calls it probably expects it to work the same way
Store Current Layout works.

Store/Restore layout submenu items need to be made checkable now
so it's easier to see which layout is the current one.

GitOrigin-RevId: 0c5ab979c56de5e1249eb800ad22939ea56ee691