[platform] IDEA-303803 Restore tool window sizes on Restore Default Layout
Tool window size was not restored if the same tool window was visible
prior to restoring layout. This was caused by the simple fact that there was no
code that would handle this situation.

Since it's not easy to write correct code that would do exactly this, considering
all possible situations, we use a simple trick here: instead of just resizing
the relevant tool window, we simply re-show it if its weight is changed and
it's visible in the new layout. Since the show code already sets the right size,
we achieve the desired effect without risking to break anything by adding
new untested code.

GitOrigin-RevId: 8e96054ea6d830db26a9b059b7e99d1ccd151102