[platform] IDEA-303803 Restore Default Layout: fix restoring split window size
The previous dirty fix didn't work for this case. If there's a split, then it would bug out,
leading to both windows hidden, but their buttons showing otherwise. This is apparently
due to trying to show windows already shown.

Fix this by doing this right this time: instead of re-showing the window, just resize its
docking area. The relevant code was extracted from ToolWindowPane.setComponent(),
now both setComponent() and the new setWeight() method delegate to the same private
method. The logic itself is left untouched, so there's a high chance that it'll continue to work.

GitOrigin-RevId: 63b954f800bd09c24a3bf89177fae62ab48ce4c2