[platform] IDEA-290231 Fix right margin color when switching themes
If the max_line_length is set in .editorconfig, then EditorSettingsManager
would tell the editor to use either null or the color from the global color scheme
to paint the right margin. This is incorrect, as the global scheme may be different
from the one applied to the editor.

Fix by simply passing AbstractColorsScheme.INHERITED_COLOR_MARKER,
which reverts the color back to the (correct) default.

GitOrigin-RevId: 8a508bee0da6e287e5f8876560c88b6338215a16