[platform] IDEA-253220 Optimize clipping when painting background image
A part of the reason why the UI with a background image is sometimes so sluggish
is because EDT spent a lot of time intersecting different clipping shapes in
IdeBackgroundUtil.MyGraphics.calcTempClip(), e.g., a Rectangle2D with
a RoundRectangle2D.

Fix by creating a temporary Graphics2D and leaving clip intersection to it by
calling clip() instead of setClip(). It uses better optimized intersection algorithms,
especially when intersecting a rectangle with something.

Now tool window resizing is no longer sluggish on my laptop's 1080p screen
with 125% scaling. It is still sluggish on the 4K screen with 150% scaling, though,
so there's definitely more to this issue than just this.

GitOrigin-RevId: 0b4a80a023fc3e91962bc332ac51556eb46fc68a