PY-53047 Exclude known reST and Epydoc tags from Grazie grammar checks
Namely, ignore everything starting with a known tag and until the closing
colon. It should match constructs like ":return:", as well as
":param list[int] x:". Excluding just the tags is not enough,
as the following identifier and colon are still recognized as part of
a grammatically-incorrect sentence.

In other languages, where markup elements in documentation comments have
their own PSI elements, this exclusion is easier. Maybe we should do the
same for reStructuredText directives at some point.

IJ-MR-96806

GitOrigin-RevId: 7e34291ff419349bb2d7c90c905fd4a1a65c3187