IDEA-297797 add option to delete old caches asynchronously during build
By default, the IDE deletes old caches synchronously that makes the build process twice longer (8 vs 4 minutes). The downside is that the deletion might not be completed if the build process is fast enough. In this case, garbage will pile up under 'system/plugins'. That's why this option is disabled by default.

Additionally, we try to use rsync on *nix systems that works faster than Java NIO or even rm. When the build process finishes, rsync continues working as its process is inherited by the root process. This might not be expected by allows to make things done.

GitOrigin-RevId: f864b0b703a898bec2267e2d19cb5536483c89c0