[platform] IDEA-253220 Round background image size to avoid scaling
Drawing images with fractional scaling is complicated and results in scaling
if the source size doesn't match the target size exactly. However, due to scaling,
it may be off by fractions of a pixel, say, 3828.25 instead of 3828 pixels. This
causes scaling, which is absolutely unneeded because it'll end up being 3828
pixels wide anyway.

Fix by slightly adjusting sizes in the user space, so they match an integer
boundary within the device space. In the worst case, it'll cause clipping the
image by one pixel, but that's acceptable.

This doesn't seem to improve performance much, though, probably because
scaling is optimized inside the JRE, so it may even detect such edge cases
and avoid actual scaling altogether, but it's better not to rely on this behavior.

GitOrigin-RevId: 2d05e5d256910ab8af8ae6f34818ff479cb245cc