[platform] IDEA-253220 Remove all HiDPI and scaling stuff when loading background image
First of all, it's impossible to scale it correctly because it could end up rendering on different
screens with different scaling factors. Also, we scale it later anyway when we do actual painting,
so it's pointless to do it beforehand (and could potentially cause quality degradation if some
scaling is actually performed when it's not needed).

GitOrigin-RevId: 290e01b604f71ae1e8ad7b0c37d8e1a4471039ce