[platform UI] IDEA-304431 Ignore a transparent vertical scrollbar width for rendering purposes by default
It used to be ignored only on Mac, where scrollbars aren't transparent anyway.

However, on non-Mac systems it led to elements being truncated by the vertical scrollbar for no good reason
with no way to read the full text except by enlarging the component or by shrinking it to the point where
the horizontal scrollbar would appear.

Fix by simply ignoring the width of the vertical scrollbar. Now the items are rendered through it regardless
of whether the horizontal scrollbar is present.

GitOrigin-RevId: 528b3213cedc16b53033dea70e92c7e14619fb43