Allow multiple ItemPresentations for each type
Fix for https://youtrack.jetbrains.com/issue/IDEA-304621

ItemPresentationProviders currently assumes that there is only a single
ItemPresentationProvider present for a given type; if that provider
returns null, it does not fall back on any further providers.

This means that plugins are not able to override any built-in behavior
for item presentation without duplicating all the logic for a given
type. (For example, changing presentation for specific Kotlin functions,
but not all Kotlin functions.) Worse, if two pluging try to override the
same type, only one of them will take effect.

This change iterates through all the providers for a given type and
returns the first non-null result. This allows multiple plugins to
augment the same type as need be, and prevents the necessity of those
plugins mimicking the built-in platform behavior for that same type.