IDEA-299807 don't re-throw manually thrown PCE from old extensions
There might be a race between NBRA listeners, when there is a pending WA,
and the listener which should've cancelled the current Job is not yet run,
but the implementation uses `runInReadActionWithWriteActionPriority` and throws PCE manually if failed.
Next `ProgressManager.checkCanceled()` will throw a correct `JobCanceledException`,
which will be handled by suspending RA, and won't be logged.

This happens in deprecated `com.intellij.ide.util.treeView.AbstractTreeUi#calculateYieldingToWriteAction`.

GitOrigin-RevId: cff4278e28a4119cd378fccafd7f5dd092843891