[kotlin] Allow multiple KotlinIconProviders
DeclarationPresenters.kt uses the IconProvider extension point to get
icons for KotlinDefaultNamedDeclarationPresentation. Currently, it finds
the first provider that is a KotlinIconProviders and uses it regardless
of whether it returns a null icon or not.

This change will move on to check for more KotlinIconProviders in the
null case, allowing plugins to register their own provider while still
having fallback to the default provider.

Note that this is the same thing that's done in the other places that
IconProvider.EXTENSION_POINT_NAME is used. So the standard practice is
try multiple providers until a non-null icon is found.

#KTIJ-23431 Fixed

GitOrigin-RevId: 7b7829f671fbe4b2df0997c44be9772baed112a0