[newUI] IDEA-303113 Debug tool window: don't recreate header actions
The reason of buttons flickering, tooltips not appearing and clicks not registering
was that these buttons were constantly recreated. The reason of that was that
some of their corresponding actions were recreated on every getChildren() call
(implemented in the local variable `toolbar` in ToolWindowHeader's initializer).

This eventually led to this check in ActionToolbarImpl.canUpdateActions() triggering:
!newVisibleActions.equals(myVisibleActions)
which would lead to rebuild of the toolbar. This happens pretty often, basically on
every action update and also every 500 ms from some timer.

Since there is no point in creating new action instances every time, the fix is simple:
just create them once and return the same instance. Then the list of actions stays
the same, unless something is really changed somewhere, and tool window header
buttons stay the same as well.

There were two culprits: RunTab.RunTabSupplier.getContentActions()
(debug window specific) and SingleContentLayout.CloseCurrentContentAction
(more general, could lead to the same issue in other tool windows). Both are now
stored in a field/property and reused.

GitOrigin-RevId: c3d33ab73508829a712f5c3b8acb22d39152b23b