introduce speed-search data-keys for BGT actions
Also, all `TextEditorComponentAction` selection actions (Ctrl-A, Ctrl-Shift-Arrows, etc.) and `SmartSelect`/`SelectWordAtCaretAction` (Ctrl-W) now work on speed search and not on the target component. Selection in tree and table speed-searches is not painted but really works. That's why the behavior is unified. We can change it in the future.

EA-705086 - REWA: ApplicationImpl.throwThreadAccessException

GitOrigin-RevId: 3309a899b3cff53b320741c6e60c5e9dbc2dda82