[newUI] EA-726044 Make PopupInlineActionsSupport accept nullable values
Some models use null values. Whether it's a good practice is doubtful, but
there is nothing in PopupInlineActionsSupport code that would not allow
nullable values, so we might as well accept them.

If it causes NPEs further down the call chain, at least we can then figure out
a more specific place that is actually at fault here. As it stands now, we have
no idea where that null comes from, because all we have in the stack trace
are generic painting code (WideSelectionListUI and related stuff).

GitOrigin-RevId: 6f2e3a1695a2955defbddf4a9eb7ce1bf8729375