[kotlin] RemoveExplicitTypeArgumentsInspection: don't suggest for inline function with reified type parameter
For such functions, type arguments are often effectively part of the name,
and the call will read awkwardly without them.

The one exception I have found in the Kotlin stdlib is `kotlin.arrayOf`:
it is an inlined reified function, but the type arguments are not significant.
Also, it is generated by J2K with explicit type arguments, so we want to remove them.

^KTIJ-1218 Fixed

GitOrigin-RevId: e82baebff99544777c503750840a3de6eaa01d16