IDEA-299725 Do not show features in productivity guide if corresponding tip not found
It will allow to move common features to the common place and do not duplicate their declaration in the different IDEs. It was not possible before, because each feature by default is showing in the guide, and it requires the tip to exist. And if there is no tip in this IDE, but feature exist, the placeholder will be shown, that is not good. But now such features will not be listed in the guide at all.

(cherry picked from commit f50682c1aa4fba30a44c54cc4ed944a44301ce65)

IJ-CR-97715

GitOrigin-RevId: 45e238ec551cacd4d8cf4c8528931ea625da9b6e