[platform] IDEA-304948 Fix setLayout() corrupting new layout
Whenever ToolWindowManagerImpl.setLayout() needed to change some
tool window's anchor, it does so via ToolWindowManagerImpl.setToolWindowAnchorImpl(),
which in turn calls ToolWindowManagerImpl.doSetAnchor(). This method has some logic
for updating other tool widows' state because their order on the strip may need to be updated.

This logic is really needed in case a tool window is moved to another anchor or when tool
windows are reordered, say, with drag and drop. However, when setting a completely new
layout, it leads to all kinds of issues because this logic tries to update the NEW layout,
which isn't completely applied yet.

This isn't easy to fix due to very complicated logic here and all kinds of cases. The safest
fix I could think of is to simply disable this logic in case the doSetAnchor() method is called
from setLayout(). To do that, add a new parameter to that method containing a reference to
the layout that needs to be updated. When it's called from setLayout(), this parameter
is set to null, which means "don't update." Seems to be working.

GitOrigin-RevId: 9cd54eb722744c30974036188aad6567bae8b661