[platform] IDEA-275323 Fix auto scroll from source interfering with mouse operations
It was fixed before (IDEA-250423) by blocking auto scroll from source while a mouse
button is pressed on the tree when it's not focused. However, that fix assumed it's
safe to resume auto scrolling once the button is released, which isn't the case because
when
a) this mouse click is a popup trigger (e.g. right mouse button on Windows),
b) AND the same mouse click causes the IDE window to become focused
(e.g. when the currently focused window is a detached editor),
then there's a race condition: focusing the IDE window causes its editor to
become focused too, which triggers auto scroll from source, and at the
same time the popup is opened, grabbing the focus. Depending on whether
the mouse button is released before or after the editor triggers auto scroll,
the tree selection may or may not jump.

Fix by not unblocking auto scroll from source on mouse release. Instead,
unblock it 500 ms after it was blocked. This is normally enough to ensure
that the editor attempts to trigger auto scroll before it's unblocked. If the system
is unusually slow it may not be enough, but in such a case a minor UI glitch is
acceptable. In addition to the timer, we also explicitly unblock auto scroll from
source whenever the popup menu is hidden. This is to make sure that if the editor
becomes focused after that, then auto scroll works as expected. Otherwise, if the
user right-clicks and immediately (within 500 ms) presses Esc, the selection would
be out of sync with the expected auto scroll behavior.

GitOrigin-RevId: be45947c61ef328ae9c4ed0afde0ce360d002531